filtros y arreglos

+++ components/ProfileScreen.js
@@ -1,6 +1,18 @@
-import { useState, useEffect } from 'react';
-import { View, Text, TouchableOpacity, StyleSheet, Image, TextInput, Button, Alert, Platform } from 'react-native';
-import { doc, setDoc, getDoc, collection } from 'firebase/firestore';
+import React, { useState, useEffect } from 'react';
+import {
+  View,
+  Text,
+  TouchableOpacity,
+  StyleSheet,
+  Image,
+  TextInput,
+  Button,
+  Alert,
+  Platform,
+  Modal,
+  Picker,
+} from 'react-native';
+import { doc, setDoc, getDoc } from 'firebase/firestore';
 import { sendPasswordResetEmail, deleteUser, updateProfile } from 'firebase/auth';
 import { auth, storage, db } from '../firebaseConfig';
 import * as ImagePicker from 'expo-image-picker';
@@ -13,6 +25,9 @@ export default function ProfileScreen({ navigation }) {
   const [dateOfBirth, setDateOfBirth] = useState('');
   const [nationality, setNationality] = useState('');
   const [image, setImage] = useState(null);
+  const [confirmModalVisible, setConfirmModalVisible] = useState(false);
+  const [actionType, setActionType] = useState('');
+
   const user = auth.currentUser;
 
   useEffect(() => {
@@ -48,7 +63,7 @@ export default function ProfileScreen({ navigation }) {
     try {
       if (user) {
         const userDoc = doc(db, 'users', user.uid);
-        await setDoc(userDoc, { name, dateOfBirth, nationality}, { merge: true });
+        await setDoc(userDoc, { name, dateOfBirth, nationality }, { merge: true });
         setIsEditing(false);
         Alert.alert('Perfil actualizado correctamente');
       }
@@ -58,26 +73,35 @@ export default function ProfileScreen({ navigation }) {
   };
 
   const handleResetPassword = async () => {
-    try {
-      if (user) {
-        await sendPasswordResetEmail(auth, user.email);
-        alert('Se ha enviado un correo electrónico para restablecer la contraseña');
-      }
-    } catch (error) {
-      console.log('Error al restablecer la contraseña', error);
-    }
+    setConfirmModalVisible(true);
+    setActionType('resetPassword');
   };
 
   const handleDeleteAccount = async () => {
+    setConfirmModalVisible(true);
+    setActionType('deleteAccount');
+  };
+
+  const confirmAction = async () => {
     try {
       if (user) {
-        await deleteUser(user);
-        navigation.navigate('LoginScreen');
-        alert('Cuenta eliminada correctamente');
+        if (actionType === 'resetPassword') {
+          await sendPasswordResetEmail(auth, user.email);
+          Alert.alert('Se ha enviado un correo electrónico para restablecer la contraseña');
+        } else if (actionType === 'deleteAccount') {
+          await deleteUser(user);
+          navigation.navigate('LoginScreen');
+          Alert.alert('Cuenta eliminada correctamente');
+        }
       }
     } catch (error) {
-      console.log('Error al eliminar la cuenta', error);
+      console.log('Error al realizar la acción', error);
     }
+    setConfirmModalVisible(false);
+  };
+
+  const cancelAction = () => {
+    setConfirmModalVisible(false);
   };
 
   const handleChooseProfileImage = async () => {
@@ -109,17 +133,17 @@ export default function ProfileScreen({ navigation }) {
     try {
       const filename = uri.substring(uri.lastIndexOf('/') + 1);
       const uploadUri = Platform.OS === 'web' ? uri.replace('file://', '') : uri;
-  
+
       const storageRef = ref(storage, `profileImages/${user.uid}/${filename}`);
       const response = await fetch(uploadUri);
       const blob = await response.blob();
-  
+
       const snapshot = await uploadBytes(storageRef, blob);
       const downloadURL = await getDownloadURL(snapshot.ref);
-      
+
       setUserProfile((prevProfile) => ({ ...prevProfile, photoURL: downloadURL }));
       const userDoc = doc(db, 'users', user.uid);
-      await setDoc(userDoc, { photoURL: downloadURL}, { merge: true });
+      await setDoc(userDoc, { photoURL: downloadURL }, { merge: true });
     } catch (error) {
       console.log('Error al cargar la imagen de perfil', error);
     }
@@ -138,34 +162,74 @@ export default function ProfileScreen({ navigation }) {
           )}
         </View>
       </TouchableOpacity>
-      <TextInput
-        style={styles.input}
-        placeholder="Nombre"
-        value={name}
-        onChangeText={(text) => setName(text)}
-        editable={isEditing}
-      />
-      <TextInput
-        style={styles.input}
-        placeholder="Fecha de Nacimiento"
-        value={dateOfBirth}
-        onChangeText={(text) => setDateOfBirth(text)}
-        editable={isEditing}
-      />
-      <TextInput
-        style={styles.input}
-        placeholder="Nacionalidad"
-        value={nationality}
-        onChangeText={(text) => setNationality(text)}
-        editable={isEditing}
-      />
+      <View style={styles.infoContainer}>
+        <Text style={styles.label}>Name</Text>
+        <TextInput
+          style={styles.input}
+          placeholder="Name"
+          value={name}
+          onChangeText={(text) => setName(text)}
+          editable={isEditing}
+        />
+      </View>
+      <View style={styles.infoContainer}>
+        <Text style={styles.label}>Birthday</Text>
+        <TextInput
+          style={styles.input}
+          placeholder="Birthday"
+          value={dateOfBirth}
+          onChangeText={(text) => setDateOfBirth(text)}
+          editable={isEditing}
+        />
+      </View>
+      <View style={styles.infoContainer}>
+        <Text style={styles.label}>Nationality</Text>
+        <Picker
+          selectedValue={nationality}
+          onValueChange={(itemValue) => setNationality(itemValue)}
+          enabled={isEditing}
+        >
+          <Picker.Item label="Select your nationality" value="" />
+          <Picker.Item label="Spain" value="Spain" />
+          <Picker.Item label="France" value="France" />
+          <Picker.Item label="Germany" value="Germany" />
+          {/* Agrega más elementos según las nacionalidades que desees */}
+        </Picker>
+      </View>
       {!isEditing ? (
-        <Button title="Editar Perfil" onPress={handleEditProfile} />
+        <Button title="Edit profile" onPress={handleEditProfile} color="green" />
       ) : (
-        <Button title="Guardar Perfil" onPress={handleSaveProfile} />
+        <Button title="Save profile" onPress={handleSaveProfile} color="green" />
       )}
-      <Button title="Restablecer Contraseña" onPress={handleResetPassword} />
-      <Button title="Eliminar Cuenta" onPress={handleDeleteAccount} />
+      <View style={styles.buttonsContainer}>
+        <Button
+          title="Reset password"
+          onPress={handleResetPassword}
+          color="blue"
+          style={styles.resetButton}
+        />
+        <Button
+          title="Delete account"
+          onPress={handleDeleteAccount}
+          color="red"
+          style={styles.deleteButton}
+        />
+      </View>
+      <Modal visible={confirmModalVisible} animationType="slide" transparent={true}>
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <Text style={styles.modalText}>
+              {actionType === 'resetPassword'
+                ? 'Are you sure that you want to reset your password?. It will send a message to your account email'
+                : 'Are you sure that you want to delete your account?'}
+            </Text>
+            <View style={styles.modalButtonsContainer}>
+              <Button title="Confirm" onPress={confirmAction} color="#dc3545" />
+              <Button title="Cancel" onPress={cancelAction} color="#f0f0f0" />
+            </View>
+          </View>
+        </View>
+      </Modal>
     </View>
   );
 }
@@ -180,7 +244,6 @@ const styles = StyleSheet.create({
     width: 150,
     height: 150,
     borderRadius: 75,
-    backgroundColor: '#ccc',
     overflow: 'hidden',
     marginBottom: 20,
   },
@@ -189,20 +252,62 @@ const styles = StyleSheet.create({
     height: '100%',
   },
   emptyProfileImage: {
-    flex: 1,
-    alignItems: 'center',
+    flex: 2,
     justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#ccc',
   },
   emptyProfileImageText: {
     fontSize: 16,
-    fontWeight: 'bold',
+    color: '#666',
   },
   input: {
     width: '80%',
     height: 40,
+    marginBottom: 20,
+    borderColor: 'gray',
     borderWidth: 1,
-    borderColor: '#ccc',
-    marginBottom: 10,
-    padding: 10,
+    borderRadius: 5,
+    paddingHorizontal: 10,
+  },
+  label: {
+    marginBottom: 5,
+    fontWeight: 'bold',
+  },
+  infoContainer: {
+    width: '80%',
+  },
+  buttonsContainer: {
+    flexDirection: 'row',
+    marginTop: 20,
+    justifyContent: 'space-around',
+    width: '80%',
+  },
+  resetButton: {
+    backgroundColor: 'blue',
+  },
+  deleteButton: {
+    backgroundColor: 'red',
+  },
+  modalContainer: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+  },
+  modalContent: {
+    backgroundColor: '#fff',
+    paddingHorizontal: 20,
+    paddingVertical: 30,
+    borderRadius: 10,
+    width: '80%',
+  },
+  modalText: {
+    fontSize: 16,
+    marginBottom: 20,
+  },
+  modalButtonsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
   },
 });
